+++ src/bot/main.py
@@ -13,7 +13,7 @@ from telegram.ext import (
 )
 from telegram.ext.filters import PHOTO, StatusUpdate
 
-from src.api.routers.hello import TELEGRAM_WEBHOOK_ENDPOINT
+from src.api.routers.telegram_webhook import TELEGRAM_WEBHOOK_ENDPOINT
 from src.bot.handlers import photo_handler, start, web_app_data
 from src.bot.jobs import send_daily_task_job, send_no_report_reminder_job
 from src.core.settings import settings
